requirements.txt
Flask==1.0.2
gcsfs==0.2.0
gunicorn==19.9.0
google-cloud-storage==1.10.0
numpy==1.15.0
pandas==0.23.4
scipy==1.1.0
scikit-learn==0.20.2
dask[complete]==1.0.0
